Break login_signup helpers into more generic files
